Refactor: Move required options a single component
